+++ Modules/Payroll/Database/Migrations/2023_01_01_071222_fix_currency_id_column_in_salary_slip_table.php
@@ -12,6 +12,9 @@ return new class extends Migration
      */
     public function up()
     {
+        if (! Schema::hasTable('salary_slips')) {
+            return;
+        }
         Schema::table('salary_slips', function (Blueprint $table) {
             $table->dropForeign(['currency_id']);
             $table->foreign('currency_id')->references('id')->on('currencies')->onUpdate('cascade')->onDelete('cascade');
